requirements.txt
discord.py
python-dotenv
numpy
pandas
scikit-learn
requests
aiohttp
aiosqlite
asyncpg